requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
python-dotenv==0.20.0  # Updated from dotenv==0.0.5
Flask==2.3.2  # Updated from Flask==0.10.1
gnureadline==6.3.3
ipdb==0.13.9  # Updated to a more recent version
ipython==7.34.0  # Updated to a more recent version
ipython-genutils==0.2.0  # Updated to a more recent version
itsdangerous==2.1.2  # Updated for compatibility with Flask 2.x
Jinja2==3.1.2  # Updated for compatibility with Flask 2.x
MarkupSafe==2.1.3  # Updated for compatibility with Jinja2 and Flask
oauthlib==3.2.1  # Updated for better compatibility
pathlib2==2.3.6  # This might not be necessary on Python 3.8+ (pathlib is part of the standard library)
pexpect==4.8.0  # Updated to a more recent version
pickleshare==0.7.5  # Updated to a more recent version
ptyprocess==0.7.0  # Updated to a more recent version
requests==2.28.2  # Updated for better security and compatibility
requests-oauthlib==1.3.1  # Updated for compatibility with requests
simplegeneric==0.8.1
six==1.16.0  # Updated to a more recent version
traitlets==5.9.0  # Updated to be compatible with ipython
tweepy==4.10.1  # Updated to a more recent version of tweepy
Werkzeug==2.3.3  # Updated for compatibility with Flask 2.x
